+++ TwilioLogic/Models/Page.cs
@@ -5,7 +5,7 @@ namespace TwilioLogic.Models
     public class Page<T>
     {
 
-        public Page(long currentPage, long totalItems, long itemsPerPage, List<T> items)
+        public Page(int currentPage, int totalItems, int itemsPerPage, List<T> items)
         {
             CurrentPage = currentPage;
             TotalPages = ((totalItems % itemsPerPage) == 0) ? (totalItems / itemsPerPage) : ((totalItems / itemsPerPage) + 1);
@@ -14,10 +14,10 @@ namespace TwilioLogic.Models
             Items = items;
         }
 
-        public long CurrentPage { get; }
-        public long TotalPages { get; }
-        public long TotalItems { get; }
-        public long ItemsPerPage { get; }
+        public int CurrentPage { get; }
+        public int TotalPages { get; }
+        public int TotalItems { get; }
+        public int ItemsPerPage { get; }
         public List<T> Items { get; }
     }
 }
